Refactor: menuController 오타 수정
